Job API: standardize methods that create actual test suites
We currently support two ways of creating test suites, because
depending on the runners they will contain different objects.
This refactor just creates two methods with a similar signature,
one for each type of runner.

One of the side effects is that it makes it clear and apparent that
the "ignore missing references" feature is missing for the resolver
based test suite creation.

Signed-off-by: Cleber Rosa <crosa@redhat.com>