Job: implement "ignore missing references" feature for the nrunner/resolver
This fills a gap that became apparent in the previous refactor.  One
way to test this feature, and compare runners is to do:

  $ avocado run /bin/true /dev/null

Which should fail, due to "/dev/null" not being resolved. Then:

  $ avocado run --ignore-missing-references /bin/true /dev/null

Should result in a pass.  The same behavior should be observed
with the addition of "--test-runner=nrunner" with the introduction
of these changes.

Note: one further possibility, not done here, and open to question, is
to change the loader code and the loader exception raised here and
make them the same one.  I refrained from doing this at this time,
because the message given by the loader exception is indeed limited to
the loader code, making reference to "avocado list" and "loader
plugins".  Also, i think it may impact other plugins such as
Avocado-VT, which implement custom loaders.

Signed-off-by: Cleber Rosa <crosa@redhat.com>