future.settings: added a 'reference' option to the 'config' command-line
This will add a new command "avocado config reference" that will print
on output the current configuration reference guide in rst format. And
can be used to update our documentation.  This PR also change our
documentation to include this reference. There is room for improvements,
for instance we can soon split the configuration into sections and also
give the ability to download a config file template.

Signed-off-by: Beraldo Leal <bleal@redhat.com>