Plugins: migrate disable and order configuration to the new settings
This leverages the information about the plugin types registered with
setuptools (which is mandatory and follows a pattern) to register the
name of the sections that will control the order of such plugins.

Then, together with the disable list, those are registered once at
the usual location for other core options.

Signed-off-by: Cleber Rosa <crosa@redhat.com>