Makefile: make the short commit of a more predictable size
git changes, from time to time, that is, from version to version, the
meaning of the '%h' formatting string, for short commits.  This
attempts to make it more stable.

BTW, git will use the minimum amount of charatecters needed to
uniquely identify a commit.  We hope that 9 charatecters will be OK
for now.

Signed-off-by: Cleber Rosa <crosa@redhat.com>