Travis CI: add ad hoc checks on OS X
While we don't support non-Linux environments at this time, attempting
to run on other environments can actually reveal valid bugs, because
of expectations that are not checked for.

The checks on this job are clearly different from the supported
environments checks, because it would crash and burn if a complete
`make check` was attempted here.  This shows a number of unittest
failures on this environment, which should be investigated (some were
addressed on previous commits), not because of the platform support
(as explained earlier) but because they may be real bugs.

Signed-off-by: Cleber Rosa <crosa@redhat.com>