avocado.utils.path.find_command: improve with a check for permissions
find_command() is used, on most if not all cases, to find a executable
path that will be executed.  But, it never checks if the file can be
executed, which defers the problem to a different part of the code,
usually causing confusion with "Permission Denied" or "FileNotFound"
kind of exceptions.

Signed-off-by: Cleber Rosa <crosa@redhat.com>