Rename args to config
A instance of `argparse.Namespace` has been used for a long time as
the main source of configuration, and that name is still present in
many places in the code.

We're now at the point where all configuration (but a few exceptions,
before the complete settings module is initialized and generates a
configuration) comes from the "avocado.core.future.settings" module,
and it may not even have a argparser or command line option
contributing to it.

So, let's rename those "args" occurences to "config" which is now
the unified output of the settings module.

Signed-off-by: Cleber Rosa <crosa@redhat.com>