nrunner: introduces the FancyDots() UI output matrix
This is an experiment with an alternative "task progress" by using a
matrix of dots. Because of the nrunner distributed architecture and the
current output list of tests, it is difficult to have an overview of how
much of the tasks are completed. This matrix attempts to give us this
visibility.

Signed-off-by: Beraldo Leal <bleal@redhat.com>