core/job: small re-org on the methods order before adding new ones
Although there is no convention today on the order of methods and
properties, and given that this class has many methods, this PR sorting
the methods here. It was becoming difficult to decide where to put new
methods on Job class. It seemed to me that the methods here were not
following any logic, so I tried to follow an intuitive pattern: magic
methods, private methods, properties, and public methods, all following
the alphabetical order.

Signed-off-by: Beraldo Leal <bleal@redhat.com>