Suite: allow empty references
Empty references should be allowed, because individual loaders can
actually come up with tests without references.  This is the case
for instance of the Avocado-VT loader, when using the `--vt-config`
command line option.

What shouldn't be the normal case, though, and should raise an
exception, is not having any tests at the end of the resolution/loader
execution.  Also, while it at, let's concentrate this behavior in a
single place.

Reference: https://github.com/avocado-framework/avocado/issues/3985
Signed-off-by: Cleber Rosa <crosa@redhat.com>