sysinfo: Introduce AVOCADO_SYSINFODIR
sysinfo collectibles commands generate not only outputs, but also
files. The output is stored in appropriate test sysinfo directory,
but not the files generated.

Now, with the introduction of new environment variable
AVOCADO_SYSINFODIR, the sysinfo commands can use them to place the
directory appropriately.
This new environment variable takes care of pre/post as well.

This was to handle the issue from:
https://github.com/avocado-framework/avocado/issues/3977

Signed-off-by: Narasimhan V <sim@linux.vnet.ibm.com>