Yaml To Mux: drop debug
The "avocado variants --debug -m $file" implementation is currently
broken, and has been for as long as 69.0.

The breakage is that it can not deal with non-simplistic files,
becomes a memory consumption hog, and Avocado eventually gets killed
by an OOM killer.  One way to reproduce is (BE WARNED it may hang your
system for a while):

  $ avocado variants --debug -m selftests/pre_release/tests/vmimage.py.data/variants.yml

Also, the requirement to choose different node classes at very late
moments, adds ennourmously to the code complexity.  IMO, it's better
to simplify the code at this time, instead of attempting to fix the
memory consumption issue and keep the current level of code
complexity.

Signed-off-by: Cleber Rosa <crosa@redhat.com>