Yaml to Mux: avoid copying a class
Earlier, all the custom functionality of the loader was based on the
functions that were defined within the main function that loads the
YAML.  Now, with all of those function unnested and defined at the
first level, it becomes clear that the custom loader needs custom
data, but not code.

Because of the way file including (!include) and namespace injection
(!using) is implemented, the "path" and "using" values are needed in
the loader.  Instead of copying a class definition, let's define those
values in the loader, and have the common constructors all defined
in the same place.

This also has the side effect of reducing the number of function
arguments.

Note: there is an opportunity here to change the way file inclusion
is implemented, to avoid the need to set that information in the
loader, but it's a major change and not attempted yet.

Signed-off-by: Cleber Rosa <crosa@redhat.com>