Introduce requirements checks on/by spawners
While some requirements may be performed in the local machine on some
circunstances (or not :), most should be performed with the knowledge
of the environment under which a task will be executed.

This introduces a basic interface on spawners, and one simple
implementation that looks for Avocado to be present on the
environments that the spawners know about.

It makes use of the requirements cache database, so once a requirement
is fulfilled for one environment, it's not checked anymore.

With regards to the cache database, there's one optmization still to
be made, which is to allow for negative caching for some
environments. For instance, on a container image, once we record that
a requirement is not available, for that same container image ID, (not
a tag) the requirement will never be available, unless, maybe the
logic of the check changes.

Signed-off-by: Cleber Rosa <crosa@redhat.com>