Functional selftests: reduce code duplication by introducing TestCaseTmpDir
Which is a simple class that contains setUp/tearDown methods that create
and cleanup a temporary directory for the test's usage.

It also changes into Avocado's base directory to allow for consistent
execution, which is a common pattern and ensures these tests are more
consistent among themselves.

And because the unittest import changed to not include mock, some
tests needed an explicit import of `unittest.mock`.

Signed-off-by: Cleber Rosa <crosa@redhat.com>