SRPM builds: enforce reproducible short commit IDs on COPR
On commit 26b4ed895 the generation of the short commit IDs was fixed, but I
completely forgot that COPR makes use of a different Makefile.

Hopefully this third attempt will be the lucky one.

In the future, it'd be nice if both Makefile and .copr/Makefile rules
could be shared/combined.

Signed-off-by: Cleber Rosa <crosa@redhat.com>