Variants: remove update_defaults
The `update_defaults()` interface is only implemented by the Yaml to
Mux plugin, and it was a requirement that was justified by plugins
such as the now archived "avocado_virt".

Test parameters can be provided by using the existing varianter
implementations (such as the simple dictionary based).  And for
plugins, a much better alternative for plugins is to register options.

Signed-off-by: Cleber Rosa <crosa@redhat.com>