Yaml to Mux: remove the ever present root node
Which is used to contribute to the idea of "default variant", an
instance of a MuxTreeNode() that is always present and can receive the
"default parameters".

Now it's only created if either files or injects are given, and the
internal Mux state depends only on that, and not on the behavior of
default_parameters().

Signed-off-by: Cleber Rosa <crosa@redhat.com>