selftests/functional/test_job_timeout.py: do not assume tests will run
Under the 5 seconds timeout that is specified in the job.  The best
way we can control the environment in which those tests run is to skip
them on general `make check` runs, and tag them so that they are
executed one at a time on
`selftests/pre_release/jobs/timesensitive.py`.

Fixes: https://github.com/avocado-framework/avocado/issues/3689
Signed-off-by: Cleber Rosa <crosa@redhat.com>