examples: fixed passjob.py with automatic case and added a custom case
Adding here some some possible use case of the current API. So far I
notice that we have two big cases: automatic job discovery vs
manual/custom job. I understand that there is room for improvements but
I would like to keep things consistent and intuitive here, because of
that this work is separating the logic behind the constructors and the
from_config() methods for both: Job() and TestSuite().

Signed-off-by: Beraldo Leal <bleal@redhat.com>