N(ext) Runner: move the status server out of the nrunner module
The only reason for the "avocado/core/nrunner.py" to be standalone is
to ease its deployment into an environment that contains nothing but
Python itself.  This will hopefully change in the future when we solve
the deployment problem.

For now, it's clear that at least the StatusServer does not need to
live in that module and be deployed when tests are executed in a
different environment.  The StatusServer is used solely by the
`avocado nrun` command, and thus, can live on another module.

Signed-off-by: Cleber Rosa <crosa@redhat.com>