Replay: introduce Job API based replay command
This introduces a new option for replay jobs, one that is expected to
replace the `avocado run --replay` option currently still kept
available.

This implementation has one very striking difference from the previous
one: it really runs a job with all of the source job configurations,
instead of picking individual configurations override.  That has the
obvious differences, but it's understood that it's much more honest
than the previous implementation.  The implementation itself is very
straightforward.

On top of this basic implementation, there could be future extensions
such as:

 * An universal configuration override option, that is, one that
   points to other Job configurations that should be overriden in
   the replayed job.

 * Modifying the source job configuration so that test resolution is
   not performed again, but reuses the resolutions (once also
   persisted) from the source job.  This would increase the job
   reproducibility and it's intended, say, for rerunning the same
   job on a different system with different resolver plugins or
   different tests;

 * Modifying the source job configuration so that test variants (and
   the parameters they contain) are not parsed and generated again,
   but retrieved from the source job in already processed form.  For
   example, when the original job variants plugins generated variants
   "A" and "B", but the environment replaying the job would, because
   of a different varianter version or behavior, would generated
   variants "A", "B" and "C".

Signed-off-by: Cleber Rosa <crosa@redhat.com>