Convert '--failfast' option to boolean
One small limitation introduced here is noted in the documentation,
that is, there's no longer a command line option to force the failfast
to the disabled during a replayed job, assuming the source job had
failfast enabled in the first place.

I believe removing this corner case is OK, given that the `avocado run
--replay` command is to be substituted in the very near future with a
`avocado replay` command, and the Job API should allow jobs to be
written for such corner cases.

Signed-off-by: Cleber Rosa <crosa@redhat.com>