N(ext) Runner: initial use of the StatusRepo on Jobs
Currently, the nrunner will keep track of a Task's status messages
in a list that is created for every task, in a serial way in its
main "task loop".  This won't work when tasks are executed in
parallel.

For that, the StatusRepo will be essential, and it's usage is
introduced here as early as possible.

Signed-off-by: Cleber Rosa <crosa@redhat.com>