avocado magic: introduce command to validate presence
Because a number of features and behavior will depend on the presence
of Avocado, as a whole, installed on a system, a way to assert that
Avocado is installed is needed.

One approach is to just look for an executable called "avocado", and
maybe check its version.  But to be extra sure, let's implement an
Avocado subcommand that outputs a magic number.  By comparing what
that command outputs, one can unmistakenly determine if Avocado is
installed.

Signed-off-by: Cleber Rosa <crosa@redhat.com>