nrun / Podman Spawner: move checks into the spawner itself
The `avocado nrun` plugin has a custom check for the podman spawner,
under which, if the podman binary does not seem to exist, it's not
even attempted to go any further.

This moves this custom check into the spawner itself, and makes way
for the users of spawners to pick the configured one and use them the
same way.  It also makes better use of the TaskInfo, by using the
status attribute.

Signed-off-by: Cleber Rosa <crosa@redhat.com>