StateMachine: work on the first entries in the "queues" by default
Instead of, on the last items.  The result here is that a job
execution (with tasks running in parallel) will more closely resemble
the order on the test suite (and thus on the command line).

And of course, if only one task is running at a time, than it will
match the order in the test suite, instead of the reverse order.

Signed-off-by: Cleber Rosa <crosa@redhat.com>