StateMachine: await after releasing the lock
This is a small optmization that should result in less "busy wait"
when all there's currently a maximum number of running tasks.

It basically means that the task itself yielding could end up waiting
a bit, and using less CPU, in place of all tasks trying their hardest
to run next.

Signed-off-by: Cleber Rosa <crosa@redhat.com>