StatusRepo: keep track of the latest (by time) status of a task
Because of the asynchronous nature of communication possible with the
N(ext) Runner, it's possible that messages arrive out of order.

A lot of the usage of the StatusRepo is to understand the current
progress of a task, that is, its status.  Let's keep track and allow
access to the most up to date status of a task.

Signed-off-by: Cleber Rosa <crosa@redhat.com>