Introduce LXC spawner functionality
Our container creation and overall Avocado setup relies on an external
fully customizable hook (which may even be a bash script or other
form of container recipe) with the major assumptions that all LXC
containers are identical reusable slots to spawn new tests in.

For instance, preparing an LXC container with full Avocado (VT)
dependencies installed could take 10+ minutes which is unfeasible
to do for each test. In the future we plan to look into LXC state
(snapshot) reuse to leverage this but it goes beyond the current
proposal here.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>