N(ext) Runner: use more consistent time reporting
First, by using time.monotonic(), we make sure that status messages
don't get generated with inconsistent times, that is, an earlier
"finished" message than a "running" one.

We also defer to the very last minute to time the message so that the
earlier "finished" than running example, which was actually happening,
doesn't happen.

Signed-off-by: Cleber Rosa <crosa@redhat.com>