fixed filter-by-tags* on resolver and loader
For now, this is a hot fix and it is not showing the resolution
information when using list with --resolver. IMO, test type and
resolution result are two different things and should be better
presented. This is creating a standard on the std output for both:
legacy and the new model. I would like to propose to add the extra
information about the resolution itself in future PRs. This closes

Signed-off-by: Beraldo Leal <bleal@redhat.com>