Suites: respect registered settings when creating from config
When calling `from_config()` the settings registered by the various
core options and plugins were not being respected.  This resulted in
crashes such as #4153.  A workaround was added previously in #4154,
but the core issue remained not solved.  With this fix, the workaround
can be removed.

Fixes: https://github.com/avocado-framework/avocado/issues/4157
Reference: https://github.com/avocado-framework/avocado/issues/4153
Reference: https://github.com/avocado-framework/avocado/issues/4154
Signed-off-by: Cleber Rosa <crosa@redhat.com>