core:runner: Fixed the suite name when showing results
Suite name as a prefix is only be displayed when we have multiple suites
or a custom suite with a name, otherwise we are going to remove this
from output.

Today, we already display the suite name as a prefix when using
--test-suite='nrunner'. This change will apply this to the default runner as
well. This closes #4083. I'm also renaming the default suite name here.

Signed-off-by: Beraldo Leal <bleal@redhat.com>