avocado.utils.process: restore newline on stdout/stderr/output
The FDDrainer reads data from processes and writes than to both a
regular logger, that has as '[stdout]' (or similar prefix) and to a
stream logger that writes the stdout/stderr/output files.

It looks for newlines in an attempt to present a human readable
result, but it fails to restore the newline on the
stdout/stderr/output files.  This restores it.

The adjusted test moves the newline into the expected STDOUT and
STDERR, while removing the newline added by "echo".

Fixes: https://github.com/avocado-framework/avocado/issues/4167
Signed-off-by: Cleber Rosa <crosa@redhat.com>