nrunner: use more consistent time reporting
This is a similar change to that in 1f5a3e887, but now applied to
the avocado-instrumented runner.

First, by using time.monotonic(), we make sure that status messages
don't get generated with inconsistent times, that is, an earlier
"finished" message than a "running" one.

We also defer to the very last minute to time the message so that the
earlier "finished" than running example, which was actually happening,
doesn't happen.

This fixes a bug in which the 'finished' status is produced with an
earlier time than the 'running' status, and causes the status repo
to discard this update because it's an older message.

Signed-off-by: Cleber Rosa <crosa@redhat.com>