nrunner: fix location and name of runner modules
While looking for suitable runners for a given kind, an
`avocado-runner-$kind` executable is tried out.  This is convenient
because any executable on the system could play the role of an
runner.

But, in some circunstances, those valid executables won't be in a
location that is the `PATH` environment variable.  For instance, on
non-activated Python virtual environments, a command such as:

 /path/to/venv/bin/avocado run --test-runner /path/to/test

Won't have `/path/to/venv/bin` in the `PATH`.  One can argue that this
is an uncommon deployment scenario, but it's a valid one and currently
used by projects such as QEMU.

Now, there's already support for alternative ways of finding the test
runners, and the next one attempted is looking for a
"avocado.core.nrunner_$kind" Python module.  This fixes the location
and module name, so that it's corretly found in the venv scenarios
mentioned earlier.

Signed-off-by: Cleber Rosa <crosa@redhat.com>