Add a smoke check on Travis CI and Windows
This simply guarantees that Avocado can be deployed on a Windows
system, can show the basic help and run "passtest.py" with the
nrunner.

With Python 3.9 released, and *not* available on Travis-CI yet,
this was a way to minimally test it too.

DISCLAIMER: this is not claiming to introducing support for Avocado on
Windows.  Still, the basic smokecheck is inteded to not allow
functionality to regress on this platform.

Signed-off-by: Cleber Rosa <crosa@redhat.com>