nrunner runners: use a default kind based on the executable name
This removes the requirement for a kind argument *if* the executable
name makes it clear what this runner handles.  It avoids the need
to given the runner name, and at the same time, the kind argument.

It's still necessary to have that option because some executables
(such as avocado-runner itself) can handle multiple kinds.

Signed-off-by: Cleber Rosa <crosa@redhat.com>