Test suite specific loader
We have multiple test suites inside one Job and each of them can have different
configurations from the user. Unfortunately, we have just one instance of the
test loader which is not able to deal with different configurations of test
suites. This commit creates for each test suite instance its own test loader
which will work with the right configuration.

Signed-off-by: Jan Richter <jarichte@redhat.com>