Integrate "make check" into a single Avocado job
The whole point about the multiple test suites support in a job is to
be able to combine varied project wide checks into an atomic check.

This combines the Job API features job already in place, and a new
suite that is the mapping of the avocado command line traditionally
used.

About the test runner selection: it's currently impossible to switch
the behavior of the human UI plugin between jobs, and the configuration
for the plugin is caught from the job configuration.  So, the default
runner is set to 'nrunner' so that it defines the human UI behavior,
which happens to work OK for the regular runner.

Signed-off-by: Cleber Rosa <crosa@redhat.com>