selftests/functional/test_replay.py: use the temporary job results dir
For the job results.  When the actual replay commands were called,
the default system (and not the temporary job results dir) were
being used.  This could cause all sorts of interesting things
because jobs such as "latest" could be re-run.

Now this uses a config file to always point to the temporary
job results dir.

Signed-off-by: Cleber Rosa <crosa@redhat.com>