selftests: create job's temporary directory within the results dir
The "run.keep_tmp" feature will change the temporary directory that a
job creates for itself, and will make it live under its own results
directory (so that it's kept).

This prevents the jobs that we create manually on tests to leave
temporary directories (at the temporary locations) which is something
our "make check" rules check with "selftests/check_tmp_dirs".

Signed-off-by: Cleber Rosa <crosa@redhat.com>