Update regex for cpu version of x86
The original regex could match with string with '()' in it, which
is not the expected version and will be written into host.cfg
during bootstrap. '()' in hosts.cfg  cannot be parsed which will
lead to list failure. This patch will set regex strict to a list
of options therefore could fix the problem.

Signed-off-by: Haijiao Zhao <haizhao@redhat.com>