Introduce example for "magic words" test type
This is a very simple, hello world like, implementation of a resolver.
If given the right (magic) words, it will return a resolution for the
"magic" test type.

It also includes the simplest possible runner for "magic" words.
It'll result in a successfull test result for URI "pass", and a test
failure for URI "fail".  For anything else, it'll result in an error.

Signed-off-by: Cleber Rosa <crosa@redhat.com>