nrunner: replacing poll() with communicate() in subprocess call
Not sure if this is the right solution, but subprocess library documentation
recommend to use "communicate() rather than .(stdout|stderr).read to avoid
deadlocks due to any of the other OS pipe buffers filling up and blocking the
child process". Issue #4322 might be one of those cases. This PR it is
replacing poll() with communicate() call.

Although the communicate() documentation says "The data read is buffered in
memory, so do not use this method if the data size is large or unlimited.",
IMO, this only occurs in some *huge* cases. I tested with some very large
outputs (>42k output lines) and managed to run without problems.

Signed-off-by: Beraldo Leal <bleal@redhat.com>