Status Server: reading stream sockets in chunks of 4096
When reading large amount of data asyncio can raise
asyncio.streams.LimitOverrunError exceptions. This is not only reading
in chunks but also sending upfront the data lenght on the first 8 bytes.
This will let the server be aware of how big is the data that client is
sending.  Probably we need a better designed protocol here, but this is
just a hot fix.

Signed-off-by: Beraldo Leal <bleal@redhat.com>