nrunner: raise/configure status server buffer limit
The asyncio streams library doesn't really allocate the amount set as
the limit, but will pause the receiving of data based on it.  Then,
when data is read and removed from the buffer, the transport will be
resumed and more data will be read.

But, when reading a line, the limit is used to determine the maximum
boundary of such a "message".  We can, without performance
implications, raise this limit and accept messages of a larger size.

Signed-off-by: Cleber Rosa <crosa@redhat.com>