assets: introduce get_asset_by_name()
Maybe this is the most controversial change. The current asset
implementation wasn't designed to support some new requirements.  For
instance: fetching an asset in "find_only" mode it is assuming that the
object it is already instantiated with a lot of arguments and besides
that, using self.fetch_asset() here will make some plugins trying to to
download the assets. Yes, a change on plugin could be made, but I
decided by this approach to avoid a bigger change.

IMO, the entire asset logic/plugin needs a bigger refactoring to
accommodate properly the new requirements. This fixes #4313.

This new method is using a class method that is much simpler.

Signed-off-by: Beraldo Leal <bleal@redhat.com>