Makefile: removing AVOCADO_EXTERNAL_PLUGINS var because it was dangerous
This variable was looking outside the avocado scope and this is
dangerous. If you have others projects with Makefile or setup.py one
level up of the currect project, then some targets here was assuming
that was Avocado related. IMO Avocado should not make this kind of
assumption and interfere with outside projects.

Signed-off-by: Beraldo Leal <bleal@redhat.com>