Makefile: migrate clean target to setup.py
This is the first step/experiment for moving Makefile targets to setup.py.
Although setuptools is the choice in the Python community to distribute /
compile packages, its documentation is not yet complete and even official
Python documentation still maintains the distutils documentation as a reference
for items not covered by setuptools. This is the case with clean command. This
is part of a bigger effort to make our build environments and CI consistent
regardless of where we are running the target commands. For sure that we have
room for improvements here but in this PR, I would like to focus on the
migration only.

Signed-off-by: Beraldo Leal <bleal@redhat.com>