Vmimage pre_release tests clean up
In the vmimage pre_release testing we are downloading all avocado supported
images, but after testing the we leave them inside avocado cache. The images
can consume dozens gigabytes of data which can cause a memory issues. Because
of this problem we have to remove downloaded images from cache after the
testing.

Signed-off-by: Jan Richter <jarichte@redhat.com>