Replay: avoid calling pre/post plugins twice
On commit 5414e18702 the calling of pre/post job plugins was moved to
the Job class itself, so it became something that happens
transparently when Job.run() is called.

The replay plugin should not be calling those again.  This fixes, for
instance, a duplicated "JOB TIME" line when using the human UI plugin.

Signed-off-by: Cleber Rosa <crosa@redhat.com>