avocado/utils/partition.py: create and keep track of loop device
Because for all matters, the device mounted is not the backing file
but the loop device.

The avoid the situation where it would be impossible to find the right
device, when one backing file is used by more than one loop device, we
better keep track of the loop device as it's created/assigned.

Signed-off-by: Cleber Rosa <crosa@redhat.com>