Enable pylint R0201
This check verifies whether the method is bound to instance. To maintain
the same functionality, all possible methods were changed to staticmethods.
Selftests was left out as it seems to be too restrictive and the check
was disabled locally in BaseRunner and ExtensionManager, being decided in
the future what is the best way to proceed with those.

Signed-off-by: Tiago Honorato <tiagohonorato1@gmail.com>