assets: introduce assets list subcommand
This command will allow users to browse easly cached files.  Since we
plan to change the asset logic, I'm not displaying, intentionally, the
"by_name"/"by_location" part. I would like to hide this detail from
users.  Also, it is important to notice that part of this code could
implemented inside the Asset class, but today we don't have a safe way
to convert a filename into an 'asset' object without passing by
assumptions implemented on this class.  This is part of the QEMU
requirements and it is related to #4311.

Signed-off-by: Beraldo Leal <bleal@redhat.com>