nrunner: provide a Task identifier when none is given
A task identifier is a mandatory attribute, and that fact is set clear
in the class initializer.  But, there are circunstances where users
are not interested in manually set them, but just have them being
unique.

Now, if None is given as an identifier, a (reasonbly guaranteed)
unique one will be set.  This was already how most tasks would
behave, since most are created from the utility function touched
here.  But, this makes it uniform in the class itself.

Signed-off-by: Cleber Rosa <crosa@redhat.com>