Cirrus CI: run selftests on Fedora
These tasks make use of the build requirements (most for running the
selftests) that are already coded in the Fedora upstream packages.

Then, it installs those depedencies and sets up a development
environment and runs the tests.  This should be very similar to the
steps that a new Avocado developer on a Fedora distro would go
through.

By having a pre-built container, instead of installing packages on
every job, we save time and increase consistency of results.

Signed-off-by: Cleber Rosa <crosa@redhat.com>