Decorators: support callable conditionals
Avocado's skipIf and skipUnless decorators allow a condition
to be given, but they have two limitations:

1) They are evaluated early
2) They are given no context

This expands the decorators by allowing them to be evaluated
later by accepting callables.  Additionally, the callables
are given the class as a parameter, this way they can have
more context on what's going on.

Signed-off-by: Cleber Rosa <crosa@redhat.com>