selftests/unit/test_utils_process.py: use a parent process other than 1
Some container systems will have PID 1 something that is not a init,
and that will have no active children.  In the case of Cirrus CI, the
containers, these are the processes available very early on (data from
/proc/$PID/stat):

   1 (pause) S 0 1 1 0 -1 4210944 543 0 0 0 0 1 0 0 20 0 1 0 894943 ...
   13 (cirrus-ci-agent) S 0 13 13 0 -1 1077952768 54023 0 0 0 248 37 ...
   44 (bash) S 13 44 44 0 -1 4210944 184 121 0 0 0 0 0 0 20 0 1 0 ...

PID 44 (bash) is a child of cirrus-ci-agent (PID 13), who is in turn,
*not* a children of PID 1.

So, for the test, let's get all the children of our parent instead,
which is guaranteed to have at least one child (ourselves) as the
test requires.

Signed-off-by: Cleber Rosa <crosa@redhat.com>