Suite: replace Tasks for Runnables
A Runnable is meant to contain all descriptive information about
what is to be run, while a Task is supposed to contain information
that is needed at run time while executing a runnable within the
context of a job (and its status server).

This change will keep Runnables in the test suite (which are simpler)
and will create the Tasks once they are about to be handed for
execution in a job.

This also solves the dilemma of changing a Task identifier, which
was not really a problem with the Tasks, but a problem caused by
them being created way too earlier than necessary.

Signed-off-by: Cleber Rosa <crosa@redhat.com>