assets: add --by-overall-limit to assets purge command
This will remove assets by an overall limit sorting assets by atime.
For instance it may be the case that a GitLab cache limit is 4 GiB, in
that case we can sort by last access, and remove all that exceeds 4 GiB
(that is, keep the last accessed 4 GiB worth of cached files). This
fixes #4457.

Signed-off-by: Beraldo Leal <bleal@redhat.com>