Selftests: do not limit the suffixes or locations of tests
There's no need to limit the location of suffixes of tests, as we
expect the resolver to do its job and treat files as test of the right
type.  The globbing is also not needed as the resolver will walk
through directories.

Also, this allows a reorganization of the directory structure for
those tests.

Signed-off-by: Cleber Rosa <crosa@redhat.com>