[NRunner] remove the structure of spawner.check_task_requirements
The original code made to check a task requirement is mapped in a way
that if requirements are defined in a test, those listed in the
check_task_requirements method should be in the list of requirements of
a test.

The mapping should work in the reversed way. If a requirement is defined
in the test, NRunner should be able to check if each of the requirements
are fulfilled and do not expect extra requirements, if they are not set.
This operation should happen in the Requirements Resolver when it is
implemented.

Signed-off-by: Willian Rampazzo <willianr@redhat.com>