TAP: unify runner implementation with exec/exec-test ones
The exec-test runner works by leveraging the exec runner, and doing
additional check on the last stage, basically deciding the actual
result of the test and what to report in the final status message.

The TAP runner is no different, but currently it basically
re-implements the execution part of the code.  Because this code
diverges, a bug fixed on the exec runner, supporting large outputs
that exceed the buffer sizes, where not applied to the TAP runner.

With this change, runners based on exec can choose to simply implement
the method that checks/builds the final status.

Signed-off-by: Cleber Rosa <crosa@redhat.com>