selftests/functional/test_skiptests.py: only created needed files
The current organization of the test class creates every single input
file for every single test.  All tests use either one or two files, so
there's quite a bit of waste going on.  And as the test grows, the
more waste is added.

Also, the proposed refactor will make it easy to make it even more
generalized and easier to write more tests.

Signed-off-by: Cleber Rosa <crosa@redhat.com>