[Requirements Resolver] adjust check_task_requirements on spawners
This adjusts the method check_task_requirements on process and podman
spawners to check for internal requirements a runtime task may have.

In a perfect world, the users should not worry about the Avocado core
dependencies or the need to specify an internal requirement in the test
itself as a test requirement, but we should have a way to check that kind
of requirement depending on the spawner the test runs on.

The only `core` check available right now is the Avocado software. With
the introduction of Package Requirements, the Avocado Software can be
handled as a requirement of `type="package"` if it is available in the
distro repository and if it is needed to run the test.

Let's remove the `core` verification on spawners for now and reintroduce
them later in a way that they are independent from the tests
requirements.

Signed-off-by: Willian Rampazzo <willianr@redhat.com>